accountType/account_type_serializers.py
from rest_framework import serializers
from accountType.models import AccountType


class AccountTypeSerializer(serializers.ModelSerializer):
    class Meta:
        model = AccountType
        fields = ["id", "type"]
