requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.11
djangorestframework==3.14.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
